# api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "myapp/v1/myapp.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/calculator/add": {
      "post": {
        "operationId": "Calculator_Add",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddRequest"
            }
          }
        ],
        "tags": [
          "Calculator"
        ]
      }
    },
    "/v1/calculator/hello": {
      "post": {
        "operationId": "Calculator_Hello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1HelloResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1HelloRequest"
            }
          }
        ],
        "tags": [
          "Calculator"
        ]
      }
    }
  },
  "definitions": {
    "v1AddRequest": {
      "type": "object",
      "properties": {
        "a": {
          "type": "integer",
          "format": "int32"
        },
        "b": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1AddResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1HelloRequest": {
      "type": "object"
    },
    "v1HelloResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "string"
        }
      }
    }
  }
}